bindings: split cgroup part of lxcfs into separate files
This was long overdue since the cgroup faking part is basically unused
at this point on most kernels.
